Add mail deleting

+++ application/models/mail_model.php
@@ -21,13 +21,17 @@ class Mail_model extends CI_Model {
 
     function addMail($userId) {
         $name = $this->input->post('mail_email');
-        $visibility = ($this->input->post('mail_access'))?1:2;
+        $date = date("Y-m-d H:i:s");
+        $visibility = ($this->input->post('mail_access')) ? 1 : 2;
         $data = array(
             'mail_email' => $name,
+            'mail_last_log_in' => $date,
             'mail_access' => $visibility,
             'mail_user_id' => $userId
         );
-        if($this->getMailByName($name)) return;
+        if ($this->getMailByName($name)){
+            return;
+        }
         $this->db->insert('mails', $data);
     }
 
@@ -37,17 +41,17 @@ class Mail_model extends CI_Model {
         $row = (array) $query->row();
         return $row;
     }
-    
-    public function getMailByName($name){        
+
+    public function getMailByName($name) {
         $this->db->where('mail_email', $name);
         $query = $this->db->get('mails');
         $row = (array) $query->row();
         return $row;
     }
 
-        public function getMailsByUser($userId){
+    public function getMailsByUser($userId) {
         $this->db->where('mail_user_id', $userId);
-        $query = $this->db->get('mails');        
+        $query = $this->db->get('mails');
         if ($query->num_rows() > 0) {
             $mails = array();
             foreach ($query->result() as $rows) {
@@ -61,13 +65,27 @@ class Mail_model extends CI_Model {
 
     function deleteMail($mailId) {
         $this->db->where('mail_id', $mailId);
+        $userId = $this->db->get('mails')->row()->mail_user_id;
+        $this->db->where('mail_user_id', $userId);
+        $query = $this->db->get('mails');
+        if ($query->num_rows() == 1) {
+            return 'single';
+        }
+        $this->db->trans_start();
+        $this->db->where('mail_id', $mailId);
         $this->db->delete('mails');
+        $this->db->trans_complete();
+        if ($this->db->trans_status() === false) {
+            return 'not';
+        } else {
+            return 'finish';
+        }
     }
 
     function changeVisibility($mailId) {
         $mail = $this->getMail($mailId);
         $visib = $mail['mail_access'];
-        $mail['mail_access'] = ($visib == 1) ? 2 : 1;        
+        $mail['mail_access'] = ($visib == 1) ? 2 : 1;
         $this->db->where('mail_id', $mailId);
         $this->db->update('mails', $mail);
         return $visib;
